Factored MidichloriansDataDictionary.h out of MidichloriansImpl.h.
