+++ Midichlorians/Dev/Beta/iOS/MidichloriansImpl.h
@@ -26,9 +26,9 @@
  *******************************************************************************/
 
 #ifndef __cplusplus
-// If you get this error build your Objective-C project, don't #include "MidichloriansImpl.h" from an ".m" file.
-// Use "Midichlorians.h" instead: It wraps the inner, plarform-independent, Midichlorians event tracking code,
-// and respects the right for the client code to be `.m`, not `.mm`.
+// If you get this error build your Objective-C project, #include neither this file nor "MidichloriansImpl.h".
+// Use "Midichlorians.h" instead.
+// It wraps the inner, plarform-independent, event tracking code, and allows client code to stay `.m`.
 #error  "This file is a C++ header, and it should be `#include`-d or `#import`-ed from an `.mm`, not an `.m` source file."
 #endif
 
@@ -39,181 +39,17 @@
 
 #include "../../../../Bricks/cerealize/cerealize.h"
 
-#ifndef COMPILE_MIDICHLORIANS_AS_SERVER_SIDE_CODE
-
 #import <Foundation/Foundation.h>
 #import <CoreLocation/CoreLocation.h>
 
-#endif  // COMPILE_MIDICHLORIANS_AS_SERVER_SIDE_CODE
+#ifdef CURRENT_MIDICHLORIANS_DATA_DICTIONARY_H
+#error "The 'MidichloriansDataDictionary.h' file should not be included prior to 'Midichlodians[Impl].h'."
+#endif  // CURRENT_MIDICHLORIANS_DATA_DICTIONARY_H
 
-struct MidichloriansEvent {
-  mutable std::string device_id;
-  mutable std::string client_id;
-  void SetDeviceId(const std::string& did) const { device_id = did; }
-  void SetClientId(const std::string& cid) const { client_id = cid; }
+#define COMPILE_MIDICHLORIANS_DATA_DICTIONARY_FOR_IOS_CLIENT
+#include "../MidichloriansDataDictionary.h"
 
-  template <class A>
-  void serialize(A& ar) {
-    ar(CEREAL_NVP(device_id), CEREAL_NVP(client_id));
-  }
-
-  virtual std::string EventAsString(const std::string& device_id, const std::string& client_id) const {
-    SetDeviceId(device_id);
-    SetClientId(client_id);
-    return JSON(WithBaseType<MidichloriansEvent>(*this));
-  }
-};
-
-// A helper macro to define structured events.
-
-#ifdef CURRENT_EVENT
-#error "The `CURRENT_EVENT` macro should not be defined."
-#endif
-
-#define CURRENT_EVENT(M_EVENT_CLASS_NAME, M_IMMEDIATE_BASE)                                            \
-  struct M_EVENT_CLASS_NAME;                                                                           \
-  CEREAL_REGISTER_TYPE(M_EVENT_CLASS_NAME);                                                            \
-  struct M_EVENT_CLASS_NAME##Helper : M_IMMEDIATE_BASE {                                               \
-    typedef MidichloriansEvent CEREAL_BASE_TYPE;                                                       \
-    typedef M_IMMEDIATE_BASE SUPER;                                                                    \
-    virtual std::string EventAsString(const std::string& did, const std::string& cid) const override { \
-      SetDeviceId(did);                                                                                \
-      SetClientId(cid);                                                                                \
-      return JSON(WithBaseType<MidichloriansEvent>(*this));                                            \
-    }                                                                                                  \
-    template <class A>                                                                                 \
-    void serialize(A& ar) {                                                                            \
-      SUPER::serialize(ar);                                                                            \
-    }                                                                                                  \
-  };                                                                                                   \
-  struct M_EVENT_CLASS_NAME : M_EVENT_CLASS_NAME##Helper
-
-// Generic iOS events.
-
-CURRENT_EVENT(iOSBaseEvent, MidichloriansEvent) {
-  std::string description;
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-    ar(CEREAL_NVP(description));
-  }
-  iOSBaseEvent() = default;
-  iOSBaseEvent(const std::string& d) : description(d) {}
-};
-
-CURRENT_EVENT(iOSIdentifyEvent, iOSBaseEvent) {
-  // An empty event, to catch the timestamp of the `identify()` call.
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-  }
-  iOSIdentifyEvent() = default;
-};
-
-CURRENT_EVENT(iOSDeviceInfo, iOSBaseEvent) {
-  std::map<std::string, std::string> info;
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-    ar(CEREAL_NVP(info));
-  }
-  iOSDeviceInfo() = default;
-  iOSDeviceInfo(const std::map<std::string, std::string>& info) : info(info) { description = "iPhoneInfo"; }
-};
-
-CURRENT_EVENT(iOSAppLaunchEvent, iOSBaseEvent) {
-  std::string binary_version;
-  std::string cf_version;
-  uint64_t app_install_time;
-  uint64_t app_update_time;
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-    ar(CEREAL_NVP(binary_version),
-       CEREAL_NVP(cf_version),
-       CEREAL_NVP(app_install_time),
-       CEREAL_NVP(app_update_time));
-  }
-  iOSAppLaunchEvent() = default;
-  iOSAppLaunchEvent(const std::string& cf_version, uint64_t app_install_time, uint64_t app_update_time)
-      : binary_version(__DATE__ " " __TIME__),
-        cf_version(cf_version),
-        app_install_time(app_install_time),
-        app_update_time(app_update_time) {
-    description = "*FinishLaunchingWithOptions";
-  }
-};
-
-CURRENT_EVENT(iOSFirstLaunchEvent, iOSBaseEvent) {
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-  }
-  iOSFirstLaunchEvent() = default;
-};
-
-CURRENT_EVENT(iOSFocusEvent, iOSBaseEvent) {
-  bool gained_focus;
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-    ar(CEREAL_NVP(gained_focus));
-  }
-  iOSFocusEvent() = default;
-  iOSFocusEvent(bool a, const std::string& d = "") {
-    description = d;
-    gained_focus = a;
-  }
-};
-
-CURRENT_EVENT(iOSGenericEvent, iOSBaseEvent) {
-  std::string event;
-  std::string source;
-  std::map<std::string, std::string> fields;
-  std::map<std::string, std::string> complex_fields;
-  std::set<std::string> unparsable_fields;
-  template <typename A>
-  void serialize(A & ar) {
-    SUPER::serialize(ar);
-    ar(CEREAL_NVP(event),
-       CEREAL_NVP(source),
-       CEREAL_NVP(fields),
-       CEREAL_NVP(complex_fields),
-       CEREAL_NVP(unparsable_fields));
-  }
-  iOSGenericEvent() = default;
-  iOSGenericEvent(const std::map<std::string, std::string>& dictionary) : fields(dictionary) {}
-
-#ifndef COMPILE_MIDICHLORIANS_AS_SERVER_SIDE_CODE
-  iOSGenericEvent(NSString * input_event, NSString * input_source, NSDictionary * input) {
-    // A somewhat strict yet safe way to parse dictionaries. Imperfect but works. -- D. K.
-    event = [input_event UTF8String];
-    source = [input_source UTF8String];
-    for (NSString* k in input) {
-      const char* key = (k && [k isKindOfClass:[NSString class]]) ? [k UTF8String] : nullptr;
-      if (key) {
-        NSObject* o = [input objectForKey:k];
-        NSString* v = (NSString*)o;
-        const char* value = (v && [v isKindOfClass:[NSString class]]) ? [v UTF8String] : nullptr;
-        if (value) {
-          fields[key] = value;
-        } else {
-          NSString* d = [o description];
-          const char* value = (d && [d isKindOfClass:[NSString class]]) ? [d UTF8String] : nullptr;
-          if (value) {
-            complex_fields[key] = value;
-          } else {
-            unparsable_fields.insert(key);
-          }
-        }
-      }
-    }
-  }
-#endif  // COMPILE_MIDICHLORIANS_AS_SERVER_SIDE_CODE
-};
-
-#ifndef COMPILE_MIDICHLORIANS_AS_SERVER_SIDE_CODE
- 
+// Define the iOS interface, being comfortably within a C++ header file included exclusively from an `.mm` one.
 @interface MidichloriansImpl : NSObject
 
 // To be called in application:didFinishLaunchingWithOptions:
@@ -227,7 +63,5 @@ CURRENT_EVENT(iOSGenericEvent, iOSBaseEvent) {
 + (void)identify:(NSString*)identifier;
 
 @end
-
-#endif
 
 #endif  // CURRENT_MIDICHLORIANS_IMPL_H
